[portafly] Rearranging types
* Index file should not include types, but export from their locations
