+++ portafly/src/types/utils.d.ts
@@ -1,0 +1,2 @@
+export type ArgumentsType<T extends (...args: any[]) => any> =
+  T extends (...args: infer A) => any ? A : never;
